Formatting fixes

+++ include/EXT2.h
@@ -59,6 +59,6 @@ InodeEntry getInodeData(int fd, Ext2Volume ext2, unsigned int inodeNum);
 
 int findFileInEXT2(char *fileName);
 
-int findFileExtVolume(int fd, Ext2Volume ext2, char *fileName, unsigned char *fileType, int inodeNumber);
+unsigned long findFileExtVolume(int fd, Ext2Volume ext2, char *fileName, unsigned char *fileType, int inodeNumber);
 
 #endif //SHOOTERC_EXT2_H
